added robotContainer back and cleaned
